LayoutTests/imported/w3c/web-platform-tests/user-timing/test_user_timing_exists-expected.txt

CONSOLE MESSAGE: line 58: ReferenceError: Can't find variable: wp_test
Description

This test validates that all of the methods used to interact with the User Timing API are defined.


Harness Error (FAIL), message = ReferenceError: Can't find variable: wp_test
